Move Launcher application code to separated launcher class
